@@ -1,6 +1,6 @@
 {
   "name": "canon-react",
-  "version": "0.7.1",
+  "version": "0.10.1",
   "description": "Rackspace Canon components built with React",
   "repository": {
     "type": "git",
@@ -11,6 +11,7 @@
     "build": "scripts/cibuild",
     "clean": "./node_modules/.bin/grunt clean",
     "test": "scripts/cibuild",
+    "server": "scripts/server",
     "documentation-build": "./node_modules/.bin/grunt documentation-build"
   },
   "keywords": [
@@ -31,23 +32,20 @@
   ],
   "author": "Adam Pickeral <adam.pickeral@rackspace.com>",
   "license": "MIT",
-  "browserify": {
-    "transform": [
-      "browserify-shim"
-    ]
-  },
   "browserify-shim": {
-    "react": "global:React"
+    "react": "global:React",
+    "react-dom": "global:ReactDOM"
   },
   "dependencies": {
     "babel-plugin-transform-runtime": "6.4.3",
     "babel-runtime": "6.3.19",
     "classnames": "^2.2.3",
     "tether": "^1.1.0",
-    "canon-react": "0.7.1"
+    "canon-react": "0.10.1"
   },
   "peerDependencies": {
-    "react": ">=0.13.3"
+    "react": "~0.14.8",
+    "react-dom": "~0.14.8"
   },
   "devDependencies": {
     "babel": "6.3.26",
@@ -56,12 +54,12 @@
     "babel-preset-react": "6.3.13",
     "babelify": "7.2.0",
     "browserify": "<4.0.0",
-    "browserify-shim": "3.8.9",
+    "browserify-shim": "3.8.12",
     "eslint": "^1.10.3",
     "eslint-plugin-react": "^3.15.0",
     "grunt": "~0.4.2",
     "grunt-babel": "6.0.0",
-    "grunt-browserify": "~1.3.0",
+    "grunt-browserify": "^5.0.0",
     "grunt-cli": "~0.1.13",
     "grunt-contrib-clean": "~0.5.0",
     "grunt-contrib-copy": "~0.5.0",
@@ -81,6 +79,8 @@
     "karma-react-jsx-preprocessor": "^0.1.1",
     "karma-spec-reporter": "0.0.23",
     "phantomjs": "^1.9.19",
-    "react": "~0.13.2"
+    "react": "~0.14.8",
+    "react-dom": "^0.14.8",
+    "react-addons-test-utils": "0.14.8"
   }
 }
