@@ -1,4 +1,5 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import Tether from 'tether';
 
 class DropdownTrigger extends React.Component {
@@ -59,7 +60,7 @@ class DropdownTrigger extends React.Component {
       this._tether = null;
     }
     if (this._dropdownNode) {
-      React.unmountComponentAtNode(this._containerDiv);
+      ReactDOM.unmountComponentAtNode(this._containerDiv);
       this._dropdownNode = null;
     }
   }
@@ -73,6 +74,7 @@ class DropdownTrigger extends React.Component {
     this._renderDropdown();
     this._listenToRootCloseEvents();
     this.setState({ isDropdownDisplayed: true });
+    this._tether.position();
   }
 
   _renderDropdown() {
@@ -81,11 +83,12 @@ class DropdownTrigger extends React.Component {
     dropdown = React.cloneElement(
       this.props.dropdown,
       {
-        hideCallback: this._hide.bind(this)
+        hideCallback: this._hide.bind(this),
+        tether: this.props.alignment
       }
     );
 
-    this._dropdownNode = React.render(dropdown, this._containerDiv);
+    this._dropdownNode = ReactDOM.render(dropdown, this._containerDiv);
     this._tether = this._createTether(this._getTetherConfig());
   }
 
@@ -98,12 +101,12 @@ class DropdownTrigger extends React.Component {
     let tetherConfig;
 
     tetherConfig = {
-      attachment: 'top left',
-      targetAttachment: 'bottom left'
+      attachment: `top ${this.props.alignment}`,
+      targetAttachment: `bottom ${this.props.alignment}`
     };
 
-    tetherConfig.element = React.findDOMNode(this._containerDiv);
-    tetherConfig.target = React.findDOMNode(this);
+    tetherConfig.element = ReactDOM.findDOMNode(this._containerDiv);
+    tetherConfig.target = ReactDOM.findDOMNode(this);
 
     return tetherConfig;
   }
@@ -118,7 +121,7 @@ class DropdownTrigger extends React.Component {
   }
 
   _handleDocumentClick(e) {
-    if (React.findDOMNode(this._dropdownNode).contains(e.target)) {
+    if (ReactDOM.findDOMNode(this._dropdownNode).contains(e.target)) {
       return;
     }
     this._hide();
@@ -136,7 +139,12 @@ class DropdownTrigger extends React.Component {
 }
 
 DropdownTrigger.propTypes = {
-  dropdown: React.PropTypes.element
+  dropdown: React.PropTypes.element,
+  alignment: React.PropTypes.oneOf(['left', 'right'])
+};
+
+DropdownTrigger.defaultProps = {
+  alignment: 'left'
 };
 
 export default DropdownTrigger;
