@@ -1,4 +1,5 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import Tether from 'tether';
 import PopoverBackground from './PopoverBackground';
 
@@ -54,14 +55,14 @@ class Popover extends React.Component {
       this._tether = null;
     }
     if (this._popoverNode) {
-      React.unmountComponentAtNode(this._containerDiv);
+      ReactDOM.unmountComponentAtNode(this._containerDiv);
       this._popoverNode = null;
     }
   }
 
   _hidePopoverBackgroundOverlay() {
     this._backgroundDiv.style.display = 'none';
-    React.unmountComponentAtNode(this._backgroundDiv);
+    ReactDOM.unmountComponentAtNode(this._backgroundDiv);
   }
 
   _removeDocumentListeners() {
@@ -77,7 +78,7 @@ class Popover extends React.Component {
     let popover;
 
     this._backgroundDiv.style.display = 'block';
-    React.render(<PopoverBackground onRequestClose={this.props.onRequestClose} />, this._backgroundDiv);
+    ReactDOM.render(<PopoverBackground isModal={ this.props.isModal } onRequestClose={this.props.onRequestClose} />, this._backgroundDiv);
     this._containerDiv.className += ' rs-popover';
 
     if (!this._tether) {
@@ -89,7 +90,7 @@ class Popover extends React.Component {
         placement: this.props.placement
       }
     );
-    this._popoverNode = React.render(popover, this._containerDiv);
+    this._popoverNode = ReactDOM.render(popover, this._containerDiv);
     this._tether.position();
   }
 
@@ -148,7 +149,7 @@ class Popover extends React.Component {
     if (this.props.offset) {
       tetherConfig.offset = this.props.offset;
     }
-    tetherConfig.element = React.findDOMNode(this._containerDiv);
+    tetherConfig.element = ReactDOM.findDOMNode(this._containerDiv);
     tetherConfig.target = this._getTarget();
     return tetherConfig;
   }
@@ -183,14 +184,15 @@ Popover.defaultProps = {
 
 Popover.propTypes = {
   children: React.PropTypes.element.isRequired,
-  placement: React.PropTypes.oneOf(['right', 'bottom-right', 'left', 'bottom-left', 'center']),
+  isModal: React.PropTypes.bool,
   isOpen: React.PropTypes.bool,
+  offset: React.PropTypes.string,
   onRequestClose: React.PropTypes.func.isRequired,
+  placement: React.PropTypes.oneOf(['right', 'bottom-right', 'left', 'bottom-left', 'center']),
   target: React.PropTypes.oneOfType([
     React.PropTypes.string,
     React.PropTypes.func
-  ]).isRequired,
-  offset: React.PropTypes.string
+  ]).isRequired
 };
 
 export default Popover;
