@@ -1,6 +1,7 @@
 import Button from '../transpiled/Button';
-import React from 'react/addons';
-let TestUtils = React.addons.TestUtils;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
 
 describe('Button', () => {
   let button, clickCalled;
@@ -16,12 +17,12 @@ describe('Button', () => {
   });
 
   afterEach(() => {
-    React.unmountComponentAtNode(React.findDOMNode(button).parentNode);
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(button).parentNode);
   });
 
   it('is enabled by default', () => {
     expect(button.props.enabled).toBe(true);
-    expect(React.findDOMNode(button)).not.toHaveClass('disabled');
+    expect(ReactDOM.findDOMNode(button)).not.toHaveClass('disabled');
   });
 
   it('renders a button', () => {
@@ -29,25 +30,25 @@ describe('Button', () => {
   });
 
   it('keeps the passed in classes', () => {
-    expect(React.findDOMNode(button)).toHaveClass('test-button-class');
+    expect(ReactDOM.findDOMNode(button)).toHaveClass('test-button-class');
   });
 
   it('keeps all passed in properties', () => {
-    expect(React.findDOMNode(button).id).toBe('button-id');
+    expect(ReactDOM.findDOMNode(button).id).toBe('button-id');
   });
 
   it('renders the text of the button', () => {
-    expect(React.findDOMNode(button).textContent).toBe('Button Text');
+    expect(ReactDOM.findDOMNode(button).textContent).toBe('Button Text');
   });
 
   it('executes the click function when clicked', () => {
-    TestUtils.Simulate.click(React.findDOMNode(button));
+    TestUtils.Simulate.click(ReactDOM.findDOMNode(button));
 
     expect(clickCalled).toBe(true);
   });
 
   it('is not hidden when hidden is false', () => {
-    expect(React.findDOMNode(button)).not.toHaveClass('rs-hidden');
+    expect(ReactDOM.findDOMNode(button)).not.toHaveClass('rs-hidden');
   });
 
   it('is hidden when hidden is true', () => {
@@ -55,23 +56,23 @@ describe('Button', () => {
       <Button hidden={true}>Button Text</Button>
     );
 
-    expect(React.findDOMNode(button)).toHaveClass('rs-hidden');
+    expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-hidden');
   });
 
-  describe('button types', () => {
-    const renderButton = (type) => {
+  describe('button canonStyles', () => {
+    const renderButton = (canonStyle) => {
       button = TestUtils.renderIntoDocument(
-        <Button type={type}>Button Text</Button>
+        <Button canonStyle={canonStyle}>Button Text</Button>
       );
     };
 
     it('action', () => {
       renderButton('action');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn');
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn-action');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn-action');
       // The spaces before and after are present on an action button because the cog and caret item surround the text
-      expect(React.findDOMNode(button).textContent).toBe(' Button Text ');
+      expect(ReactDOM.findDOMNode(button).textContent).toBe(' Button Text ');
       expect(TestUtils.findRenderedDOMComponentWithClass(button, 'rs-cog')).not.toBeNull();
       expect(TestUtils.findRenderedDOMComponentWithClass(button, 'rs-caret')).not.toBeNull();
     });
@@ -79,52 +80,52 @@ describe('Button', () => {
     it('primary', () => {
       renderButton('primary');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn');
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn-primary');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn-primary');
     });
 
     it('secondary', () => {
       renderButton('secondary');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn');
     });
 
     it('link', () => {
       renderButton('link');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn');
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn-link');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn-link');
     });
 
     it('login', () => {
       renderButton('login');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn');
-      expect(React.findDOMNode(button)).toHaveClass('rs-btn-login');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-btn-login');
     });
 
     it('cog', () => {
       renderButton('cog');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-cog');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-cog');
     });
 
     it('delete', () => {
       renderButton('delete');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-delete');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-delete');
     });
 
     it('edit', () => {
       renderButton('edit');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-edit');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-edit');
     });
 
     it('plus', () => {
       renderButton('plus');
 
-      expect(React.findDOMNode(button)).toHaveClass('rs-plus');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('rs-plus');
     });
   });
 
@@ -136,11 +137,11 @@ describe('Button', () => {
     });
 
     it('adds a disabled class to the button', () => {
-      expect(React.findDOMNode(button)).toHaveClass('disabled');
+      expect(ReactDOM.findDOMNode(button)).toHaveClass('disabled');
     });
 
     it('does not execute the click function when clicked', () => {
-      TestUtils.Simulate.click(React.findDOMNode(button));
+      TestUtils.Simulate.click(ReactDOM.findDOMNode(button));
 
       expect(clickCalled).toBe(false);
     });
