@@ -1,7 +1,8 @@
 import Dropdown from '../transpiled/Dropdown';
 import DropdownItem from '../transpiled/DropdownItem';
-import React from 'react/addons';
-let TestUtils = React.addons.TestUtils;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
 
 describe('Dropdown', () => {
   let dropdown, hideFunction;
@@ -15,34 +16,47 @@ describe('Dropdown', () => {
   });
 
   afterEach(() => {
-    React.unmountComponentAtNode(React.findDOMNode(dropdown).parentNode);
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(dropdown).parentNode);
   });
 
   it('renders a dropdown', () => {
-    expect(React.findDOMNode(dropdown)).toHaveClass('rs-dropdown');
+    expect(ReactDOM.findDOMNode(dropdown)).toHaveClass('rs-dropdown');
   });
 
   it('keeps the passed in classes', () => {
-    expect(React.findDOMNode(dropdown)).toHaveClass('test-dropdown-class');
+    expect(ReactDOM.findDOMNode(dropdown)).toHaveClass('test-dropdown-class');
   });
 
   it('defaults to action type', () => {
     expect(dropdown.props.type).toBe('action');
   });
 
-  it('renders children', () => {
+  describe('dropdown menu', () => {
     let menu;
 
-    menu = TestUtils.findRenderedDOMComponentWithClass(dropdown, 'rs-dropdown-menu');
+    beforeEach(() => {
+      menu = TestUtils.findRenderedDOMComponentWithClass(dropdown, 'rs-dropdown-menu');
+    });
 
-    expect(React.findDOMNode(menu).textContent).toBe('Dropdown Item...');
+    it('overrides positioning to static positioning if alignment is passed in as prop', () => {
+      dropdown = TestUtils.renderIntoDocument(
+        <Dropdown className='test-dropdown-class' alignment='right' hideCallback={hideFunction}><DropdownItem type='link'>Dropdown Item...</DropdownItem></Dropdown>
+      );
+      menu = TestUtils.findRenderedDOMComponentWithClass(dropdown, 'rs-dropdown-menu');
+
+      expect(ReactDOM.findDOMNode(menu).style.position).toBe('static');
+    });
+
+    it('renders children', () => {
+      expect(ReactDOM.findDOMNode(menu).textContent).toBe('Dropdown Item...');
+    });
   });
 
   it('passes hide callback down to dropdown items', () => {
     let childItem;
 
     childItem = TestUtils.findRenderedComponentWithType(dropdown, DropdownItem);
-    TestUtils.Simulate.click(React.findDOMNode(childItem));
+    TestUtils.Simulate.click(ReactDOM.findDOMNode(childItem));
 
     expect(hideFunction).toHaveBeenCalled();
   });
@@ -55,7 +69,7 @@ describe('Dropdown', () => {
         <Dropdown type='primary' hideCallback={hideFunction} />
       );
 
-      expect(React.findDOMNode(dropdownItem)).toHaveClass('rs-nav-item rs-dropdown rs-primary-dropdown');
+      expect(ReactDOM.findDOMNode(dropdownItem)).toHaveClass('rs-nav-item rs-dropdown rs-primary-dropdown');
     });
 
     it('utility', () => {
@@ -63,7 +77,7 @@ describe('Dropdown', () => {
         <Dropdown type='utility' hideCallback={hideFunction} />
       );
 
-      expect(React.findDOMNode(dropdownItem)).toHaveClass('rs-nav-item rs-dropdown rs-utility-dropdown');
+      expect(ReactDOM.findDOMNode(dropdownItem)).toHaveClass('rs-nav-item rs-dropdown rs-utility-dropdown');
     });
 
     it('action', () => {
@@ -71,7 +85,7 @@ describe('Dropdown', () => {
         <Dropdown type='action' hideCallback={hideFunction} />
       );
 
-      expect(React.findDOMNode(dropdownItem)).toHaveClass('rs-dropdown');
+      expect(ReactDOM.findDOMNode(dropdownItem)).toHaveClass('rs-dropdown');
     });
   });
 });
