@@ -3,33 +3,34 @@ import DropdownTrigger from '../transpiled/DropdownTrigger';
 import Button from '../transpiled/Button';
 import Dropdown from '../transpiled/Dropdown';
 import DropdownItem from '../transpiled/DropdownItem';
-import React from 'react/addons';
-let TestUtils = React.addons.TestUtils;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
 
 describe('DropdownTrigger', () => {
   let dropdownTrigger, tether, button;
 
   const hideFunction = () => { };
 
-  const renderDropdown = () => {
+  const renderDropdown = (menuAlignment) => {
     dropdownTrigger = TestUtils.renderIntoDocument(
-      <DropdownTrigger dropdown={<Dropdown hideCallback={hideFunction}><DropdownItem id='test-dropdown-item'>Hello</DropdownItem></Dropdown>}>
+      <DropdownTrigger alignment={menuAlignment} dropdown={<Dropdown hideCallback={hideFunction}><DropdownItem id='test-dropdown-item'>Hello</DropdownItem></Dropdown>}>
         <Button>Hello</Button>
       </DropdownTrigger>
     );
 
-    tether = jasmine.createSpyObj('tether', ['destroy']);
+    tether = jasmine.createSpyObj('tether', ['destroy', 'position']);
     spyOn(dropdownTrigger, '_createTether').and.returnValue(tether);
   };
 
   const clickTrigger = () => {
     button = TestUtils.findRenderedComponentWithType(dropdownTrigger, Button);
-    TestUtils.Simulate.click(React.findDOMNode(button));
+    TestUtils.Simulate.click(ReactDOM.findDOMNode(button));
   };
 
   describe('when rendering the DropdownTrigger', () => {
     afterEach(() => {
-      React.unmountComponentAtNode(React.findDOMNode(dropdownTrigger).parentNode);
+      ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(dropdownTrigger).parentNode);
     });
 
     it('renders the passed in trigger', () => {
@@ -37,7 +38,7 @@ describe('DropdownTrigger', () => {
 
       button = TestUtils.findRenderedComponentWithType(dropdownTrigger, Button);
 
-      expect(React.findDOMNode(button).textContent).toBe('Hello');
+      expect(ReactDOM.findDOMNode(button).textContent).toBe('Hello');
     });
 
     it('does not display the dropdown initially', () => {
@@ -59,23 +60,49 @@ describe('DropdownTrigger', () => {
         expect(dropdownTrigger._dropdownNode).not.toBeNull();
       });
 
-      it('renders the dropdown to the bottom of the trigger', () => {
+      it('renders the dropdown to the bottom of the trigger, tethering on the left by default', () => {
         renderDropdown();
         clickTrigger();
 
         expect(dropdownTrigger._createTether).toHaveBeenCalledWith({
-          element: React.findDOMNode(dropdownTrigger._containerDiv),
-          target: React.findDOMNode(dropdownTrigger),
+          element: ReactDOM.findDOMNode(dropdownTrigger._containerDiv),
+          target: ReactDOM.findDOMNode(dropdownTrigger),
           attachment: 'top left',
           targetAttachment: 'bottom left'
         });
+      });
+
+      it("tethers the dropdown on the right when value 'right' is passed in as alignment prop", () => {
+        renderDropdown('right');
+        clickTrigger();
+
+        expect(dropdownTrigger._createTether).toHaveBeenCalledWith({
+          element: ReactDOM.findDOMNode(dropdownTrigger._containerDiv),
+          target: ReactDOM.findDOMNode(dropdownTrigger),
+          attachment: 'top right',
+          targetAttachment: 'bottom right'
+        });
+      });
+
+      it('calls position on the tether', () => {
+        renderDropdown();
+        clickTrigger();
+
+        expect(tether.position).toHaveBeenCalled();
+      });
+
+      it('passes the alignment prop to the dropdown', () => {
+        renderDropdown();
+        clickTrigger();
+
+        expect(dropdownTrigger.props.dropdown.props.alignment).toBe('left');
       });
 
       it('hides the dropdown when the trigger is clicked again', () => {
         renderDropdown();
         clickTrigger();
 
-        TestUtils.Simulate.click(React.findDOMNode(button));
+        TestUtils.Simulate.click(ReactDOM.findDOMNode(button));
 
         expect(tether.destroy).toHaveBeenCalled();
         expect(dropdownTrigger._dropdownNode).toBeNull();
@@ -84,7 +111,7 @@ describe('DropdownTrigger', () => {
       it('hides the dropdown when pressing escape', () => {
         let keyUpEvent;
 
-        renderDropdown('right');
+        renderDropdown();
         clickTrigger();
 
         keyUpEvent = document.createEvent('CustomEvent');
@@ -97,10 +124,10 @@ describe('DropdownTrigger', () => {
       });
 
       it('hides the dropdown when clicking outside of the dropdown', () => {
-        renderDropdown('right');
+        renderDropdown();
         clickTrigger();
 
-        TestUtils.Simulate.click(React.findDOMNode(dropdownTrigger));
+        TestUtils.Simulate.click(ReactDOM.findDOMNode(dropdownTrigger));
 
         expect(tether.destroy).toHaveBeenCalled();
         expect(dropdownTrigger._dropdownNode).toBeNull();
@@ -122,7 +149,7 @@ describe('DropdownTrigger', () => {
     renderDropdown('right');
     clickTrigger();
 
-    React.unmountComponentAtNode(React.findDOMNode(dropdownTrigger).parentNode);
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(dropdownTrigger).parentNode);
 
     expect(tether.destroy).toHaveBeenCalled();
     expect(dropdownTrigger._tether).toBeNull();
