@@ -1,7 +1,8 @@
 import PopoverBackground from '../transpiled/PopoverBackground';
 
-import React from 'react/addons';
-const TestUtils = React.addons.TestUtils;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
 
 describe('PopoverBackground', () => {
   let popoverBackground, requestClose;
@@ -15,7 +16,7 @@ describe('PopoverBackground', () => {
   });
 
   afterEach(() => {
-    React.unmountComponentAtNode(React.findDOMNode(popoverBackground).parentNode);
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(popoverBackground).parentNode);
   });
 
   it('renders over the entire page to block events from other elements', () => {
@@ -30,11 +31,26 @@ describe('PopoverBackground', () => {
     expect(backgroundStyle.width).toEqual('100%');
     expect(backgroundStyle.height).toEqual('100%');
     expect(backgroundStyle.zIndex).toEqual('999');
+    expect(backgroundStyle['background-color']).toEqual('');
   });
 
   it('calls the request close callback when clicked', () => {
-    TestUtils.Simulate.click(React.findDOMNode(popoverBackground));
+    TestUtils.Simulate.click(ReactDOM.findDOMNode(popoverBackground));
 
     expect(requestClose).toHaveBeenCalled();
+  });
+
+  it('has a background color when isModal is true', () => {
+    let backgroundStyle, backgroundElement;
+
+    popoverBackground = TestUtils.renderIntoDocument(
+      <PopoverBackground isModal={true} onRequestClose={requestClose} />
+    );
+
+    backgroundElement = TestUtils.findRenderedDOMComponentWithClass(popoverBackground, 'rs-popover-background-overlay');
+    backgroundStyle = backgroundElement.getDOMNode().style;
+
+    // can't test for equality due to float errors. Ends up being something like 'rgba(0, 0, 0, 0.496094)'
+    expect(backgroundStyle['background-color']).not.toEqual('');
   });
 });
