@@ -1,7 +1,8 @@
 import PopoverBody from '../transpiled/PopoverBody';
 
-import React from 'react/addons';
-let TestUtils = React.addons.TestUtils;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
 
 describe('PopoverBody', () => {
   let popoverBody;
@@ -15,18 +16,18 @@ describe('PopoverBody', () => {
   });
 
   afterEach(() => {
-    React.unmountComponentAtNode(React.findDOMNode(popoverBody).parentNode);
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(popoverBody).parentNode);
   });
 
   it('renders a popover body', () => {
-    expect(React.findDOMNode(popoverBody)).toHaveClass('rs-popover-body');
+    expect(ReactDOM.findDOMNode(popoverBody)).toHaveClass('rs-popover-body');
   });
 
   it('adds an additional class', () => {
-    expect(React.findDOMNode(popoverBody)).toHaveClass('test-additional-class');
+    expect(ReactDOM.findDOMNode(popoverBody)).toHaveClass('test-additional-class');
   });
 
   it('renders children', () => {
-    expect(React.findDOMNode(popoverBody).textContent).toBe('Hello');
+    expect(ReactDOM.findDOMNode(popoverBody).textContent).toBe('Hello');
   });
 });
