@@ -1,12 +1,13 @@
 import Popover from '../transpiled/Popover';
 import PopoverOverlay from '../transpiled/PopoverOverlay';
-import React from 'react/addons';
-let TestUtils = React.addons.TestUtils;
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
 
 describe('Popover', () => {
   let popover, tether, requestCloseCallback, closeCallBackCalled;
 
-  const renderPopover = (placement, isOpen, useTargetCallback, offset) => {
+  const renderPopover = (placement, isOpen, useTargetCallback, offset, isModal) => {
     let target;
 
     setFixtures('<div id="content"><div id="some-element-id">The Target</div><div id="container"></div></div>');
@@ -24,8 +25,14 @@ describe('Popover', () => {
       target = 'some-element-id';
     }
 
-    popover = React.render(
-      <Popover placement={placement} isOpen={isOpen} onRequestClose={requestCloseCallback} target={target} offset={offset}>
+    popover = ReactDOM.render(
+      <Popover
+        placement={placement}
+        isOpen={isOpen}
+        onRequestClose={requestCloseCallback}
+        target={target}
+        offset={offset}
+        isModal={isModal}>
         <PopoverOverlay>
           This is the popover content
         </PopoverOverlay>
@@ -39,7 +46,7 @@ describe('Popover', () => {
   });
 
   afterEach(() => {
-    React.unmountComponentAtNode(document.getElementById('container'));
+    ReactDOM.unmountComponentAtNode(document.getElementById('container'));
     jasmine.getFixtures().cleanUp();
   });
 
@@ -72,8 +79,8 @@ describe('Popover', () => {
       renderPopover('right', true);
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'top left',
         targetAttachment: 'middle right',
         offset: '38px -20px'
@@ -84,8 +91,8 @@ describe('Popover', () => {
       renderPopover('bottom-right', true);
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'top left',
         targetAttachment: 'bottom right',
         offset: '-20px 45px'
@@ -96,8 +103,8 @@ describe('Popover', () => {
       renderPopover('left', true);
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'top right',
         targetAttachment: 'middle left',
         offset: '38px 20px'
@@ -108,8 +115,8 @@ describe('Popover', () => {
       renderPopover('bottom-left', true);
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'top right',
         targetAttachment: 'bottom left',
         offset: '-20px -45px'
@@ -120,8 +127,8 @@ describe('Popover', () => {
       renderPopover('bottom-left', true, true);
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'top right',
         targetAttachment: 'bottom left',
         offset: '-20px -45px'
@@ -132,8 +139,8 @@ describe('Popover', () => {
       renderPopover('bottom-left', true, false, '10px 10px');
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'top right',
         targetAttachment: 'bottom left',
         offset: '10px 10px'
@@ -144,8 +151,8 @@ describe('Popover', () => {
       renderPopover('center', true);
 
       expect(Popover.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(popover._containerDiv),
-        target: React.findDOMNode(document.getElementById('some-element-id')),
+        element: ReactDOM.findDOMNode(popover._containerDiv),
+        target: ReactDOM.findDOMNode(document.getElementById('some-element-id')),
         attachment: 'middle center',
         targetAttachment: 'middle center',
         targetModifier: 'visible'
@@ -179,11 +186,29 @@ describe('Popover', () => {
   it('cleans up when the component is unmounted', () => {
     renderPopover('right', true);
 
-    React.unmountComponentAtNode(React.findDOMNode(document.getElementById('container')));
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(document.getElementById('container')));
 
     expect(tether.destroy).toHaveBeenCalled();
     expect(popover._tether).toBeNull();
     expect(popover._popoverNode).toBeNull();
     expect(document.body.getElementsByClassName('rs-popover').length).toEqual(0);
+  });
+
+  it('renders PopoverBackground with a background color if isModal is true', () => {
+    let backgroundStyle;
+
+    renderPopover('center', true, false, '', true);
+    backgroundStyle = document.getElementsByClassName('rs-popover-background-overlay')[0].style;
+
+    expect(backgroundStyle['background-color']).not.toEqual('');
+  });
+
+  it('renders PopoverBackground with no background color if isModal is false', () => {
+    let backgroundStyle;
+
+    renderPopover('center', true, false, '', false);
+    backgroundStyle = document.getElementsByClassName('rs-popover-background-overlay')[0].style;
+
+    expect(backgroundStyle['background-color']).toEqual('');
   });
 });
