@@ -1,6 +1,7 @@
-import React from 'react/addons';
+import React from 'react';
+import ReactDOM from 'react-dom';
 import TooltipTrigger from '../transpiled/TooltipTrigger';
-let TestUtils = React.addons.TestUtils;
+import TestUtils from 'react-addons-test-utils';
 
 describe('TooltipTrigger', () => {
   let trigger, tether, triggerParent;
@@ -14,7 +15,7 @@ describe('TooltipTrigger', () => {
         <span>Tooltip Target</span>
       </TooltipTrigger>
     );
-    triggerParent = React.findDOMNode(trigger).parentNode;
+    triggerParent = ReactDOM.findDOMNode(trigger).parentNode;
   };
 
   beforeEach(() => {
@@ -23,7 +24,7 @@ describe('TooltipTrigger', () => {
 
   afterEach(() => {
     jasmine.clock().uninstall();
-    React.unmountComponentAtNode(triggerParent);
+    ReactDOM.unmountComponentAtNode(triggerParent);
   });
 
   it('renders the target', () => {
@@ -142,8 +143,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('left');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'middle right',
         targetAttachment: 'middle left',
         targetModifier: 'visible',
@@ -157,8 +158,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('bottom-left');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'top right',
         targetAttachment: 'bottom left',
         targetModifier: 'visible',
@@ -172,8 +173,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('top-left');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'bottom right',
         targetAttachment: 'top left',
         targetModifier: 'visible',
@@ -187,8 +188,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('top');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'bottom middle',
         targetAttachment: 'top middle',
         targetModifier: 'visible',
@@ -202,8 +203,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('bottom');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'top middle',
         targetAttachment: 'bottom middle',
         targetModifier: 'visible',
@@ -217,8 +218,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('right');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'middle left',
         targetAttachment: 'middle right',
         targetModifier: 'visible',
@@ -232,8 +233,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('top-right');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'bottom left',
         targetAttachment: 'top right',
         targetModifier: 'visible',
@@ -247,8 +248,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover('bottom-right');
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'top left',
         targetAttachment: 'bottom right',
         targetModifier: 'visible',
@@ -262,8 +263,8 @@ describe('TooltipTrigger', () => {
       renderTriggerAndHover();
 
       expect(TooltipTrigger.prototype._createTether).toHaveBeenCalledWith({
-        element: React.findDOMNode(document.querySelector('.rs-tooltip')),
-        target: React.findDOMNode(target),
+        element: ReactDOM.findDOMNode(document.querySelector('.rs-tooltip')),
+        target: ReactDOM.findDOMNode(target),
         attachment: 'top left',
         targetAttachment: 'bottom right',
         targetModifier: 'visible',
@@ -282,7 +283,7 @@ describe('TooltipTrigger', () => {
     TestUtils.Simulate.mouseOver(target);
     jasmine.clock().tick(250);
 
-    React.unmountComponentAtNode(React.findDOMNode(trigger).parentNode);
+    ReactDOM.unmountComponentAtNode(ReactDOM.findDOMNode(trigger).parentNode);
 
     expect(tether.destroy).toHaveBeenCalled();
     expect(trigger._tether).toBeNull();
